+++ tests/testthat/tests.brmsfit-methods.R
@@ -163,7 +163,7 @@ test_that("all S3 methods have reasonable ouputs", {
   
   # hypothesis
   hyp <- hypothesis(fit1, c("Intercept > Trt1", "Trt1:Age = -1"))
-  expect_equal(dim(hyp$hypothesis), c(2, 7))
+  expect_equal(dim(hyp$hypothesis), c(2, 8))
   expect_output(print(hyp), "(Intercept)-(Trt1) > 0", fixed = TRUE)
   expect_true(is(plot(hyp, plot = FALSE)[[1]], "ggplot"))
   
@@ -173,7 +173,7 @@ test_that("all S3 methods have reasonable ouputs", {
   expect_true(is(plot(hyp, ignore_prior = TRUE, plot = FALSE)[[1]], "ggplot"))
   
   hyp <- hypothesis(fit1, "0 > r_visit[4,Intercept]", class = "", alpha = 0.01)
-  expect_equal(dim(hyp$hypothesis), c(1, 7))
+  expect_equal(dim(hyp$hypothesis), c(1, 8))
   expect_output(print(hyp, chars = NULL), "r_visit[4,Intercept]", fixed = TRUE)
   expect_output(print(hyp), "99%-CI", fixed = TRUE)
   
@@ -181,7 +181,7 @@ test_that("all S3 methods have reasonable ouputs", {
     fit1, c("Intercept = 0", "Intercept + exp(Trt1) = 0"),
     group = "visit", scope = "coef"
   )
-  expect_equal(dim(hyp$hypothesis), c(8, 8))
+  expect_equal(dim(hyp$hypothesis), c(8, 9))
   expect_equal(hyp$hypothesis$Group[1], "1")
   
   expect_error(hypothesis(fit1, "Intercept > x"), fixed = TRUE,
@@ -197,9 +197,9 @@ test_that("all S3 methods have reasonable ouputs", {
   
   # test hypothesis.default method
   hyp <- hypothesis(as.data.frame(fit3), "bsp_meAgeAgeSD > sigma")
-  expect_equal(dim(hyp$hypothesis), c(1, 7))
+  expect_equal(dim(hyp$hypothesis), c(1, 8))
   hyp <- hypothesis(fit3$fit, "bsp_meAgeAgeSD > sigma")
-  expect_equal(dim(hyp$hypothesis), c(1, 7))
+  expect_equal(dim(hyp$hypothesis), c(1, 8))
   
   # omit launch_shiny